requirements.txt
langchain
langchain-google-genai
langchain-core
fastapi
uvicorn
twilio
openai
python-decouple
sqlalchemy
psycopg2-binary
python-multipart